Add dbeventcount verification unit test and fix proto bug (#694)
AB#1574154

Creates a unit test to check that the count of events (per event type) that we log is as expected per the data in the .xlg file. Also fixed a bug in our protobuf key since default values are not serialized in protobuf (by design) so if we had a key of value 0, it would not be serialized (and prefixmatch for this would then just match everything). As per @dannyvv we made the default enum value (0) to be invalid for "ExecutionEventId".