[macOS] compute job accounting info (#701)
This PR adds job accounting information for process pips when executing on macOS.

This info includes processor system and user time, as well as peak memory usage.

This is achieved by probing the process usage information every 2 seconds and aggregating the results.

Computing this information is currently disabled by default, and can be enabled by passing /kextMeasureProcessCpuTimes. This default is likely to change soon.