Implement proactive copy (#705)
Proactive copy has been implemented. When a piece of content is put into the distributed cache, it checks if it is the first one with this piece of content. If it is, it does a proactive copy asynchronously.

Implementation:

Machine A selects a random MachineLocation (Machine B) from already known locations.
Machine A sends machine B a gRPC message requesting that it copies a file from machine A.
Machine B will then use whatever FileCopier implementation it is using to copy the file from machine A, via a Pin.